fix usize issue

+++ src/two_d_arrays/diagonal_sum.rs
@@ -36,7 +36,7 @@ pub fn optimize_approach(arr: Vec<Vec<i32>>, rows: i32, col: i32) -> i32 {
         max_sum = max_sum + arr[i as usize][i as usize];
 
         if i != n as i32 -1-i {
-            max_sum = max_sum + arr[i as usize][n - 1 - i];
+            max_sum = max_sum + arr[i as usize][n - 1 - i as usize];
         }
     }
